+++ cocoa-beans-common/src/main/java/net/apartium/cocoabeans/structs/MinecraftVersion.java
@@ -123,6 +123,7 @@ public record MinecraftVersion(
     public static final MinecraftVersion V26_1 = new MinecraftVersion(26, 1, 0, 775);
     public static final MinecraftVersion V26_1_1 = new MinecraftVersion(26, 1, 1, 775);
     public static final MinecraftVersion V26_1_2 = new MinecraftVersion(26, 1, 2, 775);
+    public static final MinecraftVersion V26_2 = new MinecraftVersion(26, 2, 0, 776);
 
     public static final List<MinecraftVersion> KNOWN_VERSIONS = List.of(
             // 1.8 - 1.8.9
@@ -226,7 +227,10 @@ public record MinecraftVersion(
             // 26.1 - 26.1.2
             V26_1,
             V26_1_1,
-            V26_1_2
+            V26_1_2,
+
+            // 26.2
+            V26_2
     );
 
     /**